SSE2 warning regarding .NET 5.0 and .NET Core 3.1 and 2.1
Today, we've tested installing .NET 5.0 Desktop Runtime x86 for Windows 7 on our old PC with AMD Athlon XP and found that it won't install because of 0xc000001d error (Illegal instruction). That means at least a 2.00 GHz CPU with SSE2 (Athlon 64, Pentium 4) is required to run any .NET 5.0 application and that's not our fault.